+++ srcCpp03/perftest_cpp.h
@@ -86,7 +86,7 @@ class perftest_cpp
     int  _LatencyCount;
     int  _NumSubscribers;
     int  _NumPublishers;
-    int _InstanceCount;
+    unsigned long _InstanceCount;
     IMessaging *_MessagingImpl;
     char **_MessagingArgv;
     int _MessagingArgc;
